reservation-api/requirements.txt
aniso8601==9.0.1
Babel==2.9.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Babel==2.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.26
protobuf==3.18.1
PyMySQL==1.0.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25
Werkzeug==2.0.2